Move mixins into base k8s classes
